+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support was added for the [Svelte language](https://svelte.dev/)
   ([#1805](https://github.com/ianlewis/todos/issues/1805)).
+- A new `--no-error-on-unsupported` flag was added to suppress errors when
+  unsupported files are passed via the command line
+  ([#1751](https://github.com/ianlewis/todos/issues/1751)).
 
 ### Changed